move carousel to product list and version bumped branch to 1.3.1
