+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "agroeasy",
-  "version": "1.3.0",
+  "version": "1.3.1",
   "description": "",
   "main": "server.js",
   "scripts": {
