+++ src/components/home/components/Home.js
@@ -3,7 +3,6 @@ import { Col, Row } from 'antd';
 
 import Search from './Search';
 import  productList  from '../../productList';
-import CarouselImages  from './Carousel';
 import { HOME_STRINGS } from '../constants';
 
 const { ProductList } = productList.components;
@@ -27,7 +26,6 @@ class Home extends React.Component {
                         </Col>
                     </Row>
                 </div>
-                <CarouselImages />
                 <ProductList />
             </div>
         );
