+++ src/components/home/constants.js
@@ -22,17 +22,3 @@ export const HOME_STRINGS = {
         patronize us. Give us a try.`,
 };
 
-export const CAROUSEL_DATA = [
-    {
-        src: '/carousel-placeholder-1.png',
-        text: 'This is Image 1',
-    },
-    {
-        src: '/carousel-placeholder-1.png',
-        text: 'This is Image 2',
-    },
-    {
-        src: '/carousel-placeholder-1.png',
-        text: 'This is Image 3',
-    },
-];
