+++ src/components/productList/components/styles.css
@@ -1,8 +1,40 @@
 .product-list {
-    padding-top: 20px;
+  padding-top: 20px;
 }
 
 .card-image {
-    width: 240;
-    height: 300px;
+  width: 240;
+  height: 300px;
+}
+
+/* carousel.js */
+.ant-carousel {
+  padding: 20px 0;
+}
+
+.ant-carousel .slick-slide {
+  text-align: center;
+  overflow: hidden;
+}
+
+.ant-carousel .slick-slide img {
+  height: 400px;
+  width: 100%;
+}
+
+.ant-carousel .slick-slide .carousel-text {
+  position: absolute;
+  bottom: 150px;
+  width: 100%;
+}
+
+.ant-carousel .slick-slide h3 {
+  color: #fff;
+  text-transform: uppercase;
+  font-size: 40px;
+  font-family: monospace;
+}
+
+.ant-carousel .slick-dots {
+  bottom: 40px;
 }
